Fix route setup naming.
I think this was a leftover from the example. 
It would conflict with anyone that has a `Moderate` operation and used that name in the setup routes function. 

BTW, thanks for sharing this, I will test it later and let you know if I found something that can be improved/fixed :)

Best,
Pedro